+++ test/test-runner/test-run-watch-create.mjs
@@ -1,0 +1,8 @@
+// Test run({ watch: true }) runs new tests when a new file is created in the watched directory
+import '../common/index.mjs';
+import { skipIfNoWatch, refreshForTestRunnerWatch, testRunnerWatch } from '../common/watch.js';
+
+skipIfNoWatch();
+refreshForTestRunnerWatch();
+
+await testRunnerWatch({ action: 'create', fileToCreate: 'new-test-file.test.js', useRunApi: true });
