+++ test/test-runner/test-run-watch-delete.mjs
@@ -1,0 +1,8 @@
+// Test run({ watch: true }) does not throw when deleting a watched test file
+import '../common/index.mjs';
+import { skipIfNoWatch, refreshForTestRunnerWatch, testRunnerWatch } from '../common/watch.js';
+
+skipIfNoWatch();
+refreshForTestRunnerWatch();
+
+await testRunnerWatch({ fileToUpdate: 'test.js', action: 'delete', useRunApi: true });
