+++ test/test-runner/test-run-watch-repeatedly.mjs
@@ -1,0 +1,8 @@
+// Test run({ watch: true }) runs tests repeatedly
+import '../common/index.mjs';
+import { skipIfNoWatch, refreshForTestRunnerWatch, testRunnerWatch } from '../common/watch.js';
+
+skipIfNoWatch();
+refreshForTestRunnerWatch();
+
+await testRunnerWatch({ file: 'test.js', fileToUpdate: 'test.js', useRunApi: true });
