+++ test/test-runner/test-run-watch-without-file.mjs
@@ -1,0 +1,8 @@
+// Test run({ watch: true }) supports running tests without a file
+import '../common/index.mjs';
+import { skipIfNoWatch, refreshForTestRunnerWatch, testRunnerWatch } from '../common/watch.js';
+
+skipIfNoWatch();
+refreshForTestRunnerWatch();
+
+await testRunnerWatch({ fileToUpdate: 'test.js', useRunApi: true });
